Ignore subtrees that do not contain any possible requires.
These changes make AST traversal virtually free, leaving parsing as the
dominant cost of what detective does.

The algorithm uses a regular expression to find all possible substrings of
interest, then uses the indexes of those matches to guide the traversal of
the AST, typically ignoring most of the tree.